Do not kill processes on CI using clean task
Leaking processes are now killed by dedicated CI steps before and after each build. This change allows to remove the workaround for clean task ordering.
